refactored tests to new file
